openai-agents: adopt the reviewer's patch from fed2c6a
Takes fed2c6a wholesale: the instrumentor now builds its own
TelemetryHandler instead of taking the get_telemetry_handler singleton,
so instrument() honors the providers it is handed. That is the root
cause of the test-isolation problem the previous commit worked around
in conftest, so the reset fixture is gone.

Also switches to the reviewer's cassette and moves the end-to-end
failing-tool test next to it in test_instrumentor.py.

Restores pytest.register_assert_rewrite ahead of the vcr import; the
reordering in that patch left it after the module was already imported
and raised PytestAssertRewriteWarning on every run.
