+++ instrumentation/opentelemetry-instrumentation-genai-openai-agents/tests/conftest.py
@@ -3,7 +3,7 @@
 
 from __future__ import annotations
 
-from collections.abc import Iterator
+import os
 
 import pytest
 
@@ -12,12 +12,17 @@ pytest.register_assert_rewrite("opentelemetry.test_util_genai.vcr")
 from opentelemetry.test_util_genai.vcr import (
     scrub_response_headers_overwrite,
 )
-from opentelemetry.util.genai.handler import get_telemetry_handler
 
 pytest_plugins = [
     "opentelemetry.test_util_genai.fixtures",
     "opentelemetry.test_util_genai.vcr",
 ]
+
+
+@pytest.fixture(autouse=True)
+def environment():
+    if not os.getenv("OPENAI_API_KEY"):
+        os.environ["OPENAI_API_KEY"] = "test_openai_api_key"
 
 
 @pytest.fixture(scope="module")
@@ -38,25 +43,3 @@ def vcr_config():
             }
         ),
     }
-
-
-@pytest.fixture
-def reset_telemetry_handler() -> Iterator[None]:
-    """Drop util-genai's process-wide ``TelemetryHandler`` around each test.
-
-    ``get_telemetry_handler`` caches the first handler it builds. A test that
-    instruments without explicit providers would otherwise pin every later
-    test to the global ones, so their in-memory exporters would come back
-    empty no matter what the instrumentation did.
-    """
-
-    def _clear() -> None:
-        if (
-            getattr(get_telemetry_handler, "_default_handler", None)
-            is not None
-        ):
-            delattr(get_telemetry_handler, "_default_handler")
-
-    _clear()
-    yield
-    _clear()
